requirements.txt
flask
tensorflow
pandas
numpy
scikit-learn
neptune-tensorflow-keras
neptune-client